+++ src/tissuemoisture.h
@@ -17,7 +17,7 @@ double segmentRelativeWaterContent(double psiSym, double pi0, double epsilon,
                                   double af);
 double tissueRelativeWaterContent(double psiSym, double pi0, double epsilon, 
                                   double psiApo, double c, double d, 
-                                  double af, double femb = 0.0);
+                                  double af);
 
 double turgorLossPoint(double pi0, double epsilon);